+++ CRM/Civirules/Upgrader.php
@@ -774,5 +774,11 @@ class CRM_Civirules_Upgrader extends CRM_Civirules_Upgrader_Base {
     return TRUE;
   }
 
+  public function upgrade_2065() {
+    $this->ctx->log->info('Applying update 2065 - Adding condition Contact has not an active membership of type.');
+    CRM_Civirules_Utils_Upgrader::insertConditionsFromJson($this->extensionDir . DIRECTORY_SEPARATOR . 'sql/conditions.json');
+    return TRUE;
+  }
+
 }